+++ src/domain/index.js
@@ -1,0 +1,4 @@
+export { default as Lotto } from './Lotto/Lotto.js';
+export { default as LottoBuyer } from './LottoBuyer/LottoBuyer.js';
+export { default as RateOfReturnCalculator } from './RateOfReturnCalculator/RateOfReturnCalculator.js';
+export { default as WinningRank } from './WinningRank/WinningRank.js';
