+++ src/views/OutputView.js
@@ -1,6 +1,6 @@
-import Console from '../utils/console.js';
-
 import { FORMAT_MESSAGE, OUTPUT_MESSAGE } from '../constants/messages/messages.js';
+
+import Console from '../utils/console.js';
 
 const OutputView = Object.freeze({
   printLottoCount(lottoCount) {
